Add files via upload
This is the script itself. At this point, it must be used via the command line, though I intend to create a GUI for it at some point. Please feel free to suggest features and improvements, and of course this script is available for use by anyone who so desires it.