Update adiabatic.rb
Minor refactoring: converted standard if-then structure to ternary (for readability) in the #parcel_is_saturated? and #stable? methods. 

Logic changes: Having issues with cloud_thickness accuracy, so reduced by one (fixed 'off by one' error). To be honest, I'm not sure this did it. It's been a while since I took this class, and I don't recall the exact criteria. Cloud thickness can still be determined using the stability and saturation readings for the time being.